# apps/server/docs/swagger-spec.yaml
openapi: 3.0.0
info:
  title: Shopify Partner CRM API
  description: API for managing Shopify partner CRM operations including partner accounts, organizations, leads, and activities.
  version: '1.0.0'
  contact:
    name: API Support
    url: https://api.cartrabbit.com/support
    email: support@cartrabbit.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://example.com/terms
...
paths:
...
  /organization:
    post:
      summary: Create an organization
      tags:
        - Organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: 'humbletree'
                userId:
                  type: string
                  example: 'ck2j49r2a0d1m0741gzhjo41v'
                role:
                  type: string
                  enum:
                    - ADMIN
                    - MEMBER
              required:
                - name
                - userId
      responses:
        '201':
          description: Organization created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Organization created successfully.'
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 'org12345'
                      name:
                        type: string
                        example: 'humbletree'
                      userId:
                        type: string
                        example: 'ck2j49r2a0d1m0741gzhjo41v'
                      createdAt:
                        type: integer
                        format: int64
                        example: 872748344
                      updatedAt:
                        type: integer
                        format: int64
                        example: 124234123
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Invalid data for organization. Please check the provided data.'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Unauthorized access. Please provide a valid API key.'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'An unexpected error occurred while creating the organization.'
  /organization/{userId}:
    get:
      summary: Get all organizations by user ID
      tags:
        - Organization
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            example: 'ck2j49r2a0d1m0741gzhjo41v'
      responses:
        '200':
          description: Organizations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Organizations retrieved successfully.'
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: 'org12345'
                        name:
                          type: string
                          example: 'humbletree'
                        userId:
                          type: string
                          example: 'ck2j49r2a0d1m0741gzhjo41v'
                        createdAt:
                          type: integer
                          format: int64
                          example: 872748344
                        updatedAt:
                          type: integer
                          format: int64
                          example: 124234123
                  integrationCount:
                    type: integer
                    format: int64
                    example: 2
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Invalid user ID. Please check the provided user ID.'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Unauthorized access. Please provide a valid API key.'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'No organizations found for the provided user ID.'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'An unexpected error occurred while retrieving the organizations.'
  /organization/{id}:
    delete:
      summary: Delete an organization by ID
      tags:
        - Organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 'org12345'
      responses:
        '200':
          description: Organization deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Organization deleted successfully.'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Invalid organization ID. Please check the provided ID.'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Unauthorized access. Please provide a valid API key.'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Organization not found.'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'An unexpected error occurred while deleting the organization.'

    patch:
      summary: Update an organization's name
      tags:
        - Organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 'org12345'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: 'newname'
              required:
                - name
      responses:
        '200':
          description: Organization name updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Organization name updated successfully.'
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 'org12345'
                      name:
                        type: string
                        example: 'newname'
                      userId:
                        type: string
                        example: 'ck2j49r2a0d1m0741gzhjo41v'
                      createdAt:
                        type: integer
                        format: int64
                        example: 872748344
                      updatedAt:
                        type: integer
                        format: int64
                        example: 124234123
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Invalid data for organization. Please check the provided data.'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Unauthorized access. Please provide a valid API key.'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Organization not found.'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'An unexpected error occurred while updating the organization.'
  /org-member/{memberId}:
    patch:
      summary: Update the role of a member (only admin can change this)
      tags:
        - OrgMember
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            example: 'ck2j49r2a0d1m0741gzhjo41v'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                  enum:
                    - ADMIN
                    - MEMBER
                  example: 'MEMBER'
              required:
                - role
      responses:
        '200':
          description: Member role updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Member role updated successfully.'
                  data:
                    type: object
                    properties:
                      memberId:
                        type: string
                        example: 'ck2j49r2a0d1m0741gzhjo41v'
                      role:
                        type: string
                        enum:
                          - ADMIN
                          - MEMBER
                        example: 'MEMBER'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Invalid data for role update. Please check the provided data.'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Unauthorized access. Only admin can change the role.'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Member not found.'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'An unexpected error occurred while updating the member role.'
    delete:
      summary: Delete a member by ID
      tags:
        - OrgMember
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            example: 'ck2j49r2a0d1m0741gzhjo41v'
      responses:
        '200':
          description: Member deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Member deleted successfully.'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Invalid member ID. Please check the provided ID.'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Unauthorized access. Only admin can delete a member.'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Member not found.'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'An unexpected error occurred while deleting the member.'
...